Removal of Component Module
The team decided that the Component module did not serve a desirable purpose for the evolving project. The removal of the Component module consists of the following:

Files Deleted: ccomponent.cpp, component.hpp

Files Modified: makefile, pswitch_main.cpp, packetswitch.cpp, packetswitch.hpp

The modifications above have conclude with a successful compilation and execution of the PSP program. The removal of the Component module will allows for a greater focus on the remaining modules for SQA.